🚚 :: 리소스 계층에 맞게 변경
