--- maeumgagym-core/src/main/kotlin/com/info/maeumgagym/auth/model/AccessToken.kt
@@ -1,10 +1,0 @@
-package com.info.maeumgagym.auth.model
-
-data class AccessToken(
-
-    val subject: String,
-
-    val accessToken: String,
-
-    val ttl: Long
-)
