--- maeumgagym-core/src/main/kotlin/com/info/maeumgagym/auth/model/RefreshToken.kt
@@ -1,10 +1,0 @@
-package com.info.maeumgagym.auth.model
-
-data class RefreshToken(
-
-    val subject: String,
-
-    val rfToken: String,
-
-    val ttl: Long
-)
