+++ maeumgagym-infrastructure/src/main/kotlin/com/info/maeumgagym/global/jwt/JwtResolver.kt
@@ -1,6 +1,6 @@
 package com.info.maeumgagym.global.jwt
 
-import com.info.maeumgagym.domain.auth.repository.AccessTokenRepository
+import com.info.maeumgagym.global.jwt.repository.AccessTokenRepository
 import com.info.maeumgagym.global.env.jwt.JwtProperties
 import com.info.maeumgagym.global.exception.InvalidTokenException
 import org.springframework.stereotype.Component
